Entity migration to new pack system

+++ main/java/minecrafttransportsimulator/systems/pack/PackInstrument.java
@@ -3,7 +3,7 @@ package minecrafttransportsimulator.systems.pack;
 
 public class PackInstrument {
 
-    public double[] pos;
+    public float[] pos;
     public int[] rot;
 
 }
