+++ main/java/minecrafttransportsimulator/systems/pack/PackWindow.java
@@ -3,8 +3,8 @@ package minecrafttransportsimulator.systems.pack;
 
 public class PackWindow {
 
-    public double[] x;
-    public double[] y;
-    public double[] z;
+    public float[] x;
+    public float[] y;
+    public float[] z;
 
 }
